TS-620 More 64-bit fixes to the SM
git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/trunk@1056140 13f79535-47bb-0310-9956-ffa450edef68

+++ example/append-transform/append-transform.c
@@ -108,7 +108,7 @@ handle_transform(TSCont contp)
   data = TSContDataGet(contp);
   if (!data) {
     towrite = TSVIONBytesGet(write_vio);
-    if (towrite != INT_MAX) {
+    if (towrite != INT64_MAX) {
       towrite += append_buffer_length;
     }
     data = my_data_alloc();
