+++ proxy/TestProxy.cc
@@ -177,7 +177,7 @@ struct TestProxy:Continuation
       if (!data)
         return done();
       remote = (VConnection *) data;
-      clusterOutVIO = remote->do_io(VIO::WRITE, this, INT_MAX, inbuf);
+      clusterOutVIO = remote->do_io(VIO::WRITE, this, INT64_MAX, inbuf);
       ink_assert(clusterOutVIO);
       SET_HANDLER(tunnelEvent);
       tunnel = new OneWayTunnel(remote, vc, this, TUNNEL_TILL_DONE, true, true, true);
@@ -192,7 +192,7 @@ struct TestProxy:Continuation
     if (!vc)
       return done();
     SET_HANDLER(startEvent);
-    vc->do_io(VIO::READ, this, INT_MAX, inbuf);
+    vc->do_io(VIO::READ, this, INT64_MAX, inbuf);
     return EVENT_CONT;
   }
 
@@ -289,7 +289,7 @@ struct TestProxy:Continuation
     *url_end = 0;
     sprintf(outbuf->start, "GET %s HTTP/1.0\nHost: %s\n\n", url, host);
     outbuf->fill(strlen(outbuf->start) + 1);
-    remote->do_io(VIO::WRITE, this, INT_MAX, outbuf);
+    remote->do_io(VIO::WRITE, this, INT64_MAX, outbuf);
     // printf("sending [%s]\n",outbuf->start);
     return EVENT_CONT;
   }
@@ -324,7 +324,7 @@ struct TestProxy:Continuation
     *url_end = 0;
     sprintf(outbuf->start, "GET %s HTTP/1.0\nHost: %s\n\n", url, host);
     outbuf->fill(strlen(outbuf->start) + 1);
-    remote->do_io(VIO::WRITE, this, INT_MAX, outbuf);
+    remote->do_io(VIO::WRITE, this, INT64_MAX, outbuf);
     // printf("sending [%s]\n",outbuf->start);
     return EVENT_CONT;
   }
@@ -382,7 +382,7 @@ struct TestAccept:Continuation
   {
     if (event == NET_EVENT_ACCEPT) {
       MIOBuffer *buf = new_MIOBuffer();
-        e->do_io(VIO::READ, new TestProxy(buf), INT_MAX, buf);
+        e->do_io(VIO::READ, new TestProxy(buf), INT64_MAX, buf);
     } else
     {
       printf("TestAccept error %d\n", event);
